+++ GeneHunter/Source/Monster/AffinitiesComponent/Public/SAffinity.h
@@ -43,7 +43,7 @@ public:
 	// Events
 	, _OnMinusClicked([](){return FReply::Unhandled();})
 	, _OnPlusClicked([](){return FReply::Unhandled();})
-	, _OnCircleClicked([](const int AffinityIndex, const int CurrentPoints){return FReply::Unhandled();})
+	, _OnCircleClicked([](const uint8 AffinityIndex, const uint8 CurrentPoints){return FReply::Unhandled();})
 	, _Enabled([](){return true;})
 
 		{}
