+++ GeneHunter/Source/Monster/StatsComponent/Public/Stats/StandardStat.h
@@ -11,5 +11,5 @@ struct STATSCOMPONENT_API FStandardStat : public FStat
 	GENERATED_BODY()
 	
 public:
-	virtual float CalculateValue(const int Level) override;
+	virtual float CalculateValue(const uint16 Level) override;
 };
